backend/log-service/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
httpx==0.28.1
pydantic==2.12.4
python-multipart==0.0.6
pydantic-settings==2.12.0
PyJWT==2.10.1